move the `convert_search_lines_to_bank_polygons` function to the `BankLines` class
